+++ asrGenerator/argumentparser.py
@@ -6,31 +6,32 @@ argParserStatus = Enum('ArgParserStatus', 'SUCCESS FAILURE')
 
 class Arguments:
     def __init__(self, input = None, module = None, output = None, status = argParserStatus.FAILURE):
-        self.inputFile = input
-        self.module = module
-        self.outputFolder = output
-        self.parseStatus = status
+        self.__inputFile = input
+        self.__module = module
+        self.__outputFolder = output
+        self.__parseStatus = status
     
     def get_input_file(self):
-        return self.inputFile
+        return self.__inputFile
 
     def get_module(self):
-        return self.module
+        return self.__module
 
     def get_output_folder(self):
-        return self.outputFolder
+        return self.__outputFolder
 
     def get_parser_status(self):
-        return self.parseStatus
+        return self.__parseStatus
 
 def parseArguments():
     parser = argparse.ArgumentParser()
-    parser.add_argument('-i', '--input', help = 'Absolute file path to the autosar module definition file')
-    parser.add_argument('-m', '--module', help = 'The module for which the python classes should be generated')
-    parser.add_argument('-o', '--output', help = 'Location where the python classes should be generated')
-    args = parser.parse_args()
+    parser.add_argument('-i', '--input', required = True, help = 'Absolute file path to the autosar module definition file')
+    parser.add_argument('-m', '--module', required = True, help = 'The module for which the python classes should be generated')
+    parser.add_argument('-o', '--output', required = True, help = 'Location where the python classes should be generated')
     
+    args = parser.parse_args()
     inputFile = args.input
+    
     if inputFile is None:
         logging.error('Module definition file is not provided.')
     else: